+++ makeModule.ps1
@@ -10,7 +10,7 @@ $hshModManifestParams = @{
 	#DefaultCommandPrefix = ""
 	#FormatsToProcess = "SomeModule.format.ps1xml"
 	ModuleToProcess = "vNuggletsUtilityMod.psm1"
-	ModuleVersion = "1.0.0"
+	ModuleVersion = "1.1.0"
 	## scripts (.ps1) that are listed in the NestedModules key are run in the module's session state, not in the caller's session state. To run a script in the caller's session state, list the script file name in the value of the ScriptsToProcess key in the manifest
 	NestedModules = @("vNuggletsUtilityMod_functions.ps1", "vNugglets_SupportingFunctions.ps1")
 	PowerShellVersion = [System.Version]"4.0"