+++ vNugglets.Utility/vNuggletsUtilityMod_functions.ps1
@@ -1,3 +1,66 @@
+function Find-VNVMWithDuplicateMACAddress {
+<#  .Description
+    Get information about the VM(s) that have a network adapter whose MAC address is a duplicate of another network adapter in the vCenter(s) to which this PowerCLI session is connected
+
+    .Synopsis
+    Get information about duplicate MAC addresses
+
+    .Example
+    Find-VNVMWithDuplicateMACAddress
+    VMName                DuplicatedMAC         MoRef                                                Count
+    ------                -------------         -----                                                -----
+    {myVM03, oldVM322}    00:50:56:3F:FF:FF     {VirtualMachine-vm-16277, VirtualMachine-vm-109}         2
+
+    Find VMs with network adapters whose MAC address is the same, and return a bit of info about them
+
+    .Example
+    Find-VNVMWithDuplicateMACAddress
+    VMName     DuplicatedMAC         MoRef                       Count
+    ------     -------------         -----                       -----
+    myVM21     00:50:56:00:00:09     VirtualMachine-vm-16277         2
+
+    Find VM (just one, apparently, in this vCenter) with network adapters whose MAC address is the same -- in this case, the VM has at least two network adapters, both of which have the same MAC address
+
+    .Link
+    http://vNugglets.com
+
+    .Outputs
+    System.Management.Automation.PSCustomObject
+#>
+    [CmdletBinding()]
+    [OutputType([System.Management.Automation.PSCustomObject])]
+    Param () ## end param
+
+    process {
+        ## get VirtualMachine .NET views where the items is not marked as a Template
+        $colDevMacAddrInfo = `
+        Get-View -ViewType VirtualMachine -Property Name,Config.Hardware.Device -Filter @{"Config.Template" = "False"} | Foreach-Object {
+           $viewThisVM = $_
+           $_.Config.Hardware.Device | Where-Object {$_ -is [VMware.Vim.VirtualEthernetCard]} | Foreach-Object {
+               New-Object -Type PSObject -Property @{VMName = $viewThisVM.Name; MacAddr = $_.MacAddress; MoRef = $viewThisVM.MoRef}
+           } ## end foreach-object
+        } ## end foreach-object
+
+        ## get the non-unique MAC addresses (if any),
+        $arrDuplicatedMAC_GroupInfo = $colDevMacAddrInfo | Group-Object MacAddr | Where-Object {$_.count -gt 1}
+
+        ## for each duplicated MAC, return an object with the given properties
+        if ($null -ne $arrDuplicatedMAC_GroupInfo) {
+            $arrDuplicatedMAC_GroupInfo | Foreach-Object {
+                New-Object -Type PSObject -Property ([ordered]@{
+                    VMName = $_.Group | Foreach-Object {$_.VMName} | Select-Object -Unique
+                    DuplicatedMAC = $_.Name
+                    MoRef = $_.Group | Foreach-Object {$_.MoRef} | Select-Object -Unique
+                    Count = $_.Count
+                }) ## end new-object
+            } ## end foreach-object
+        } ## end if
+        else {Write-Verbose "no duplicate MAC addresses found on non-template VMs"}
+    } ## end process
+} ## end function
+
+
+
 function Get-VNNetworkClusterInfo {
 <#  .Description
     Get information about the VMware HA Cluster(s) in which the given virtual network (virtual portgroup) is defined. May 2015
@@ -377,9 +440,9 @@ function Get-VNVMByVirtualPortGroup {
 
 function Get-VNVMEVCInfo {
 <#  .Description
-    Code to get VMs' EVC mode and that of the cluster in which the VMs reside.  May 2014, Matt Boren
+    Function to get VMs' EVC mode and that of the cluster in which the VMs reside
     .Example
-    Get-VNVMEVCInfo -Cluster myCluster | ?{$_.VMEVCMode -ne $_.ClusterEVCMode}
+    Get-Cluster myCluster | Get-VNVMEVCInfo | ?{$_.VMEVCMode -ne $_.ClusterEVCMode}
     Name        PowerState   VMEVCMode   ClusterEVCMode   ClusterName
     ----        ----------   ---------   --------------   -----------
     myvm001     poweredOff               intel-nehalem    myCluster0
@@ -388,8 +451,8 @@ function Get-VNVMEVCInfo {
     Get all VMs in given clusters where the VM's EVC mode does not match the Cluster's EVC mode
 
     .Example
-    Get-VM myVM | Get-VNVMEVCInfo
-    Get the EVC info for the given VM and the cluster in which it resides
+    Get-VM myVM0,myVM1 | Get-VNVMEVCInfo
+    Get the EVC info for the given VMs and the cluster in which they reside
 
     .Outputs
     System.Management.Automation.PSCustomObject
@@ -397,11 +460,11 @@ function Get-VNVMEVCInfo {
     [CmdletBinding(DefaultParameterSetName="ByCluster")]
     [OutputType([System.Management.Automation.PSCustomObject])]
     param(
-        ## Cluster name pattern (regex) to use for getting the clusters whose VMs to get
-        [parameter(ParameterSetName="ByCluster",Position=0)][string]$Cluster = ".+",
+        ## Cluster whose VMs about which to get EVC information
+        [parameter(ValueFromPipeline=$true,ParameterSetName="ByCluster",Position=0)][VMware.VimAutomation.Types.Cluster[]]$Cluster,
 
-        ## Id/MoRef of VM for which to get EVC info
-        [parameter(ValueFromPipelineByPropertyName=$true,ParameterSetName="ByVMId",Position=0)][Alias("Id","MoRef")][string[]]$VMId
+        ## VM for which to get EVC info
+        [parameter(ValueFromPipeline=$true,ParameterSetName="ByVM",Position=0)][VMware.VimAutomation.Types.VirtualMachine[]]$VM
     ) ## end param
 
     begin {
@@ -426,7 +489,7 @@ function Get-VNVMEVCInfo {
 
         Switch ($PSCmdlet.ParameterSetName) {
             "ByCluster" {
-                Get-View -ViewType ClusterComputeResource -Property Name,Summary -Filter @{"Name" = $Cluster} | Foreach-Object {
+                Get-View -Property Name,Summary -Id $Cluster.Id | Foreach-Object {
                     $viewThisCluster = $_
                     Get-View -ViewType VirtualMachine @hshParamForGetVMView -SearchRoot $viewThisCluster.MoRef | Foreach-Object {
                         _New-InfoObj -VMView $_ -ClusterEVCModeKey $viewThisCluster.Summary.CurrentEVCModeKey -ClusterName $viewThisCluster.Name
@@ -434,8 +497,8 @@ function Get-VNVMEVCInfo {
                 } ## end foreach-object
                 break
             } ## end case
-            "ByVMId" {
-                Get-View @hshParamForGetVMView -Id $VMId | Foreach-Object {
+            "ByVM" {
+                Get-View @hshParamForGetVMView -Id $VM.Id | Foreach-Object {
                     ## update the View data to get the cluster name and the cluster summary (which has the cluster's EVCMode)
                     $_.UpdateViewData("Runtime.Host.Parent.Name")
                     $_.Runtime.LinkedView.Host.LinkedView.Parent.UpdateViewData("Summary")